+++ client/package.json
@@ -6,6 +6,7 @@
     "@emotion/react": "^11.4.1",
     "@emotion/styled": "^11.3.0",
     "@fortawesome/fontawesome-free": "5.15.2",
+    "@lingui/react": "3.13.3",
     "@material-ui/core": "^4.12.3",
     "@mui/icons-material": "^5.0.0",
     "@mui/material": "^5.0.6",
